refactor: move document status enum to const folder
