+++ src/entities/document/const/DocumentStatus.ts
